Revamp geoaxis
Since https://github.com/JuliaPlots/GeoMakie.jl/commit/8bfe471d5dfc82d2eab590a9574811fed62fbd7b, the geoaxis has been built on top of the regular MakieLayout Axis.  This is a great idea and 
streamlines a lot of the code, but in doing so a lot of the original logic
was not reimplemented in terms of ticklabels and other things.  Currently, 
GeoAxis uses the normal Axis ticks, which are placed in screen-space, not dataspace,
meaning that they are not subject to the nonlinear transform.

This commit hooks into and hijacks some of the Axis machinery in order to allow 
the standard methods of interacting with an Axis to work with GeoAxis.  As such,
the tick finding algorithm is now linked in to the Axis; minor grids have still
not been implemented, but those are easy to implement with the current foundation.  
Visibility, color, linestyle, linewidth, etc. are all forward from the usual Axis
attributes, and the Axis default drawings are all set to be invisible.  

As of now, tick labels are close to working; there is an odd scaling factor which
I have yet to figure out.  However, after that is done, GeoMakie's functionality
should be back to (and better than) the original!

This means that zooming should now work smoothly, though I have not been able to
test it yet.  I am currently working on trying to get the labels to work; the
next feature after that will be automatic limit finding with the algorithm from 
https://github.com/JuliaPlots/GeoMakie.jl/pull/100.